decomposing Options part for widget subjects
